usb_modeswitch: don't return a value from stack
The memory is not allocated anymore when the function returns and that
causes no end of mayhem and undefined behavior.

Let's make it static so that it's perserved after the return. It wastes
LINE_DIM bytes of memory and requires the caller to consume the value
before another call, but that's no problem.

http://www.draisberghof.de/usb_modeswitch/bb/viewtopic.php?f=2&t=2557