usb_modeswitch: don't reset an endpoint unless stalled
A Nokia CS-21M-02 sometimes ignores the first bulk message after an
endpoint reset. Let's not reset the endpoint unless we know it's stalled
and rely on the kernel driver to leave the device in a sensible state after
unbinding.

It seems to depend on traffic prior to the reset attempt -- for some reason it
typically won't happen if an unit ready test precedes the reset, but would
happen after other commands, such as sense request or allowing medium removal.
Timing doesn't make a difference.