usb_modeswitch: don't pass busname to the dispatcher
In the better case it's empty and useless, in worst case it's something
very irrelevant (a bus name of a random parent of a device that happens to have
one, such as the usb controller) and harmful -- when passed to the
dispatcher it appends it to the sysfs path, making it invalid.

To add insult to injury, the / is escaped as _ for systemd (as opposed
to the correct "\x2d"), so the dispatcher gets even more confused.

Both of these issues exhibit on Odriod C2, with DWC2 platform USB
controlers, that happen to have device-tree names of "dwc2_a" and "dwc2_b".
Udev uses that for %b of the modem device:

When the actual device path is
/sys/devices/platform/dwc2_b/usb1/1-1/1-1.4/1-1.4:1.0 (and a link in
/sys/bus/usb/devices/1-1.4:1.0), udev helper starts the service with
dwc2_b_1-1.4:1.0 argument, which the dispatcher translates to
/sys/bus/usb/devices/dwc2/b.
/sys/bus/usb/devices/dwc2_b/1-1.4:1.0 would still be incorrect.

Let's just lay it to rest.